Dev feature pi12 80 - Branch API DACQ (#25)
* Dev feature pi12 11 (#6) (#13)

* Make test.vue a meteo getter for province

TernaManager etc

TernaManager2 13/06

Script Tabella TernaToken e TernaController Setup

TernaToken API, Dto, Domain, Controllers, Mapper

prima pulizia del codice

prima pulizia del codice

ripristino ProvinceController

Aggiunta script creazione database (#3)

* Aggiunto script per la creazione del DB e spostato i 3 "originali" in una sottocatella

* commit di prova da db master

* prova dopo fetch

* altra prova come le altre

* ????

* ultima prova lo giuro non picchiatemi

* Rimozione file di testo usato per i test di push

---------


Script aggiornato dopo alcune correzioni minori (#4)

Bugfix DB (#5)

* Bugfix DB

* Bugfix DB

TernaManager2 13/06

TernaToken API, Dto, Domain, Controllers, Mapper

* Debug TernaToken

* Porta 3306

---------


ApiManager, RetrieveData and ProcessConfigFile

RetrieveData fino ad estrazione Dto type da JSON config

Funzioni ApiManager, Script Tabella LogsApi, Update ApiConfig

ApiCallLogs Domain & Dto

Config classes for ApiCalls and TernaToken, UpdateTable v0.5, pre-test

If(Auth) condition, setup if(Frequency)

Update file config chiamate Api

Check API in base a frequenza update, pt 1 + Update Mapper per ApiCallsLogs

CRUD operations for ApiCallLogs

Update Jsonfile api config e relativa Dto

Update Init a Set nelle Domain Classes

API Manager: inserimento try/catch per Retrieve Data, commenti, renamed "template" --> "apiCall", check StatusCode 200 e aggiunta CRUD ApiCallsLogs in RetrieveData

RetrieveData final update pre-Testing

Creazione Ambiente Test e Program 0.1

Scripts Update

Debug RetrieveData pt1

- Risolto deserializzazione file json configurazione (con relative modifiche al file json e a Dto.JsonApiConfig
- Inserimento "configuration" come parametro ApiManager (constructor)
- Rimozione ProtocolloConfiguration (obsoleto)
- Spostamento script SQL e aggiornamento (parallelo a DB group)

Retrieve Data debug, fino ad UpdateTable

Commodity Debug for RetrieveData

Add CommodityData dto class

UpdateTable -- condizioni Commodity

filteredDataValue control works

Script SQL tabella date

Update dto and domain Date

Refactoring ApiManager function order

Update Table - Condizioni Commodity 2 (Date conditions)

Commodity Daily Api Successful Test

Commodities Historic & Daily Api calls, Tested

QoL improvements

Update scripts commodity & ternaToken

TernaToken --> Token

TernaToken Retrieve + Update, Generation Retrieve (Ongoing)

Token terna updates and is fed correctly to the api call, solved Generation API issues with returning XML instead of JSON

Generation Daily calls for all energy types, need to check how consistently they work at runtime.

Update Generation Api with Task.Delay 2s

Update api config structrue - added "lag" parameter

Dto.DailyPrice creation and Dto.Price update

api config update: added "dtoDataClass" field, and subsequent updates to dtoClasses

GetDtoList function, Price Api conditions 0.1

Dto WeatherData & DailyPrice for API data

Generation 0.5 and Weather 0.1

Weather updates, prep for rebase

* DailyPrice Test Completed + QoL post-rebase changes

* apiConfig Price Historical calls from 04/2022 to 07/2023

* Changes to Macrozone/Bidding_zone (csv and sql) to match API data

* Load API daily & history calls, still issue with string to float data conversion

* SOLVED: string to float conversion issue

* Generation History 24/07/23 to today (valid until 24/09/2023, or 60 days-ish)

* Weather API completed

* float to double to solve db issues

* Weather History calls & funcs, no params "Windspeed_80m", "snowdepth", and "shower" because the two apis don't match 100%

* Master script DB, update to 23_08, for fresh installlation (to be tested on container or fresh machine)

* Weather API update, 3 new config files (one per year) + Forecast file in attempt to speed process up. Only partial success (currently takes about 20 minutes to load 1 history API call, ca 23k lines)

* ReadMe dataset

* Weather Updates

* WeatherDaily correction

* Revert "WeatherDaily correction"

This reverts commit b88f39e3f684ba2bf4fae3917f0a826e55516048.

* API refactoring (speed improvement)

* Revert "API refactoring (speed improvement)"

This reverts commit 24c7cd826881860cb042e29e0f5d53c2d0196d62.

* Revert "Revert "API refactoring (speed improvement)""

This reverts commit 5064e05ea897687109eb2e3868b9010f4d42c032.

* appsettings 3306

+++ ProgettoIndustriale.Type/Domain/Price.cs
@@ -11,13 +11,13 @@ namespace ProgettoIndustriale.Type.Domain
         public int Id { get; set; }
 
         [Column("base_price_EURxMWh")]
-        public float BasePriceEur { get; set; }
+        public double BasePriceEur { get; set; }
 
         [Column("incentive_component_EURxMWh")]
-        public float IncentiveComponentEur { get; set; }
+        public double IncentiveComponentEur { get; set; }
 
         [Column("unbalance_price_EURxMWh")]
-        public float UnbalancePriceEur { get; set; }
+        public double UnbalancePriceEur { get; set; }
 
         [Column("COD_macrozone")]
         public int IdMacroZone { get; set; }
